fixed midi file

+++ midi.py
@@ -72,14 +72,14 @@ def samples_to_midi(samples, fname, ticks_per_sample, thresh=0.5):
     mid.tracks.append(track)
     ticks_per_beat = mid.ticks_per_beat
     ticks_per_measure = 4 * ticks_per_beat
-    ticks_per_sample = ticks_per_measure / samples_per_measure
+    ticks_per_sample = int(ticks_per_measure / samples_per_measure)
     abs_time = 0
     last_time = 0
     for sample in samples:
         for y in range(sample.shape[0]):
             abs_time += ticks_per_sample
             for x in range(sample.shape[1]):
-                note = x + (128 - num_notes)/2
+                note = int(x + (128 - num_notes)/2)
                 if sample[y,x] >= thresh and (y == 0 or sample[y-1,x] < thresh):
                     delta_time = abs_time - last_time
                     track.append(Message('note_on', note=note, velocity=127, time=delta_time))
